requirements.txt
pandas
numpy
telegram-send
dataframe-image
tqdm
fastparquet
plotly
openpyxl
xlrd
kaleido
pillow
statsmodels
linearmodels
scikit-learn
python-dotenv
tabulate
matplotlib
seaborn
